Time: 3 ms (63.82%), Space: 6 MB (10.25%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -11,7 +11,7 @@ public:
         if(rev>INT_MAX || rev<INT_MIN){
             return 0;
         }
-        return (int)rev;
+        return (int)(int)rev;
         
     }
 };
